Data.Fixed.Decimal

Description

Decimal types of fixed precision and scale which can use any Integral type to store mantissa

Documentation

module Data.Fixed.Decimal.Class

newtype Decimal p (s :: Nat)

Decimal type of fixed precision and scale which uses:

Integral p => p type to store mantissa

Type-level number s :: Nat to define scale (fractional part length)

Constructors

Decimal

Fields

mantissa :: p