Data.Fixed.Decimal.Class

Documentation

class FixedDecimal d where

Fixed decimal types class

Associated Types

type Precision d

Integral type for storing mantissa

type Scale d :: Nat

The size of the fractional part in decimal digits

Methods

scale :: d -> Int

Runtime value of Scale d

decimal :: Integral m => m -> Int -> d

Value constructor

decimal m e constructs a decimal number via m * 10 ^^ e.